A large modification post-fork.
So backstory, this website is a fork of a project I attempted to make
and failed miserably. However, a lot of code was a very good foundation
for this project. Thus, I forked it and used it here and removed
previous references to the old project.

I also started working on the basic backend functions and some API
routes, nothing major, just ‘getting started’ kind of stuff.

TL;DR - Cleanup from old project (as this is a fork), getting a little
started on things